refactor(app): lift the two biggest branches out of the live-envelope effect (STR-2)
`MqttLiveWiring.tsx` was 1,705 lines, 1,101 of them in a single `useEffect`,
which is the last of the giant files the audit named (#1228 took the other
three components and the three Rust files). Two branches of the live-envelope
handler account for most of it and come out whole:

    handle-acp-event.ts     490 lines — deltas, tool use/results, plan
                                        updates, the terminal status
    handle-live-message.ts  191 lines — message rows and the agent-reply
                                        flush path

The file goes 1,705 → 1,037, and the effect 1,101 → ~420.

Both branches always ended in a branch-level `return`, so they always ended
the envelope callback; the call sites return after them, which is exactly the
control flow they had. Their closures turned out small — 7 and 5 names from
the component body — so they take an explicit `LiveWiringContext` instead of
capturing. The dependency is unchanged; it is just written down now, and the
context is built inside the effect rather than in the component body so it
captures what the inline branches captured: the versions current when the
effect ran, not a new object each render (which would also drag the object
into the effect's dep array).

**This file has no tests** — nothing in the suite references it — so the
extraction is verified by construction rather than by behaviour: the moved
bodies are line-for-line identical to `origin/main`, compared as a multiset of
stripped lines, nothing added and nothing removed. The only edits are the
signatures, the `ctx` destructure, and one re-narrowing guard
(`if (!decoded.acpEvent) return;`) that restores what the enclosing `if` used
to give TypeScript.

eslint reports the same 3 warnings / 0 errors on this file as `main` does, so
no new lint debt; 3311 frontend tests pass.

Still large and still untested: the remaining ~420-line effect, and
`recoverStaleLiveStreams` / `flushTurnAgentReply` in the component body. Worth
characterisation tests before the next cut, not another mechanical move.
